first test1

+++ routes/story.js
@@ -2,11 +2,11 @@ const express = require("express");
 const { db } = require("../connection");
 const { ObjectId } = require("mongodb");
 const jwt = require("jsonwebtoken");
-const { cloud_uplod, cloud_remove } = require("../cloud")
+const { cloud_Multiple_uplod, cloud_remove } = require("../cloud")
 const { upload } = require("../multerfunction")
 const path = require("path")
 const fs = require("fs");
-const { date } = require("joi");
+
 
 
 const router = express.Router()
@@ -32,32 +32,67 @@ router.get("/index", async (req, res) => {
 })
 
 
-router.get("/",async(req,res)=>{
+router.get("/", async (req, res) => {
 
     const story = db.collection("story")
 
-    try{
+    try {
 
-     data = await story.aggregate([
-        {
-            $lookup:{
-          from:"user",
-          localField:"user",
-          foreignField: "_id",
-          as: "author"
-            } 
-          },
-          { $project: {  "author.password": 0, "author.cover": 0 } },
-     ]).toArray()
+        data = await story.aggregate([
+            { $sort: { date: -1 } },
+            {
+                $group: {
+                    _id: "$user",
+                    user: { $first: "$$ROOT" }
+                }
+            },
+            { $replaceRoot: { newRoot: "$user" } },
+            {
+                $lookup: {
+                    from: "user",
+                    localField: "user",
+                    foreignField: "_id",
+                    as: "author"
+                }
+            },
+            { $project: { "author.password": 0, "author.cover": 0 } },
+        ]).toArray()
 
 
-     res.status(200).json({"data":data})
+        res.status(200).json({ "data": data })
 
     }
     catch (err) {
         console.log("=========>" + err);
         res.status(500).send("err in " + err)
     }
+})
+
+router.get("/user/:id", async (req, res) => {
+
+    const story = db.collection("story")
+
+    try {
+        data = await story.aggregate([
+            { $match: { "user": new ObjectId(req.params.id) } },
+            {
+                $lookup: {
+                    from: "user",
+                    localField: "user",
+                    foreignField: "_id",
+                    as: "author"
+                }
+            },
+            { $project: { "author.password": 0, "author.cover": 0 } },
+        ]).toArray()
+
+        res.status(200).json({ "data": data })
+    }
+    catch (err) {
+        console.log("=========>" + err);
+        res.status(500).send("err in " + err)
+    }
+
 })
 
 router.post("/", async (req, res) => {
@@ -80,8 +115,8 @@ router.post("/", async (req, res) => {
     try {
 
         data = await story.insertOne({
-            "img":[],
-            "info":info,
+            "img": [],
+            "info": info,
             "paragraph": req.body.paragraph,
             "date": new Date(),
             "user": new ObjectId(req.user.id)
@@ -98,7 +133,7 @@ router.post("/", async (req, res) => {
 })
 
 
-router.post("/:id/img", upload.single("img"), async (req, res) => {
+router.post("/:id/img", upload.array("imgs"), async (req, res) => {
 
     const story = db.collection("story")
 
@@ -127,35 +162,29 @@ router.post("/:id/img", upload.single("img"), async (req, res) => {
             return res.status(400).json({ message: "dont allowed" })
         }
 
-        if (!req.file) {
-            return res.status(403).json({ message: "you not send img" })
+        const uploder = async (path) => await cloud_Multiple_uplod(path, "imges")
 
-        }
-        const pathimge = path.join(__dirname, "../upload/" + req.file.originalname)
+        const urls = []
 
-        if (livestory.img.originalname == req.file.originalname) {
-            fs.unlinkSync(pathimge)
+        const files = req.files
 
-            return res.status(200).json({ message: "upload img Succeed 1" })
-        }
 
-        result = await cloud_uplod(pathimge)
+        for (const file of files) {
 
-        if (livestory.img.publicid !== null) {
-            cloud_remove(livestory.img.publicid)
+            const { path, originalname } = file
+
+            const newpath = await uploder(path)
+
+            urls.push({ newpath, originalname })
+
+            fs.unlinkSync(path)
         }
 
         await story.updateOne({ "_id": new ObjectId(req.params.id) }, {
-            $set: {
-                "img": {
-                    "url": result.secure_url,
-                    "publicid": result.public_id,
-                    "originalname": req.file.originalname,
-                }
+            $push: {
+                "img": { $each: urls }
             }
         })
-
-        fs.unlinkSync(pathimge)
 
         res.status(200).json({ message: "upload img Succeed", })
 
@@ -168,7 +197,7 @@ router.post("/:id/img", upload.single("img"), async (req, res) => {
 })
 
 
-router.put("/:id",async(req,res)=>{
+router.put("/:id/pull/img", async (req, res) => {
 
     const story = db.collection("story")
 
@@ -182,8 +211,7 @@ router.put("/:id",async(req,res)=>{
     } else {
         return res.status(400).json({ messege: "login frist" })
     }
-
-    try{
+    try {
 
         livestory = await story.findOne({ "_id": new ObjectId(req.params.id) })
 
@@ -196,13 +224,16 @@ router.put("/:id",async(req,res)=>{
 
             return res.status(400).json({ message: "dont allowed" })
         }
+        await story.updateOne({ "_id": new ObjectId(req.params.id) }, {
+            $pull: {
+                "img": { "publicid": req.body.publicid } // publicid
 
+            }
+        })
 
-        await story.updateOne({"_id":new ObjectId(req.params.id)},{$set:{
-            "paragraph":req.body.paragraph
-        }})
+        cloud_remove(req.body.publicid)
 
-        res.status(200).json({"message": "story updated"})
+        res.status(200).json({ message: "done" })
 
     }
     catch (err) {
@@ -213,7 +244,7 @@ router.put("/:id",async(req,res)=>{
 })
 
 
-router.delete("/:id",async(req,res)=>{
+router.put("/:id", async (req, res) => {
 
     const story = db.collection("story")
 
@@ -228,7 +259,7 @@ router.delete("/:id",async(req,res)=>{
         return res.status(400).json({ messege: "login frist" })
     }
 
-    try{
+    try {
 
         livestory = await story.findOne({ "_id": new ObjectId(req.params.id) })
 
@@ -242,16 +273,63 @@ router.delete("/:id",async(req,res)=>{
             return res.status(400).json({ message: "dont allowed" })
         }
 
-        if(livestory.img.publicid != null){
 
-            cloud_remove(livestory.img.publicid )
+        await story.updateOne({ "_id": new ObjectId(req.params.id) }, {
+            $set: {
+                "info": info
+            }
+        })
+
+        res.status(200).json({ "message": "story updated" })
+
+    }
+    catch (err) {
+        console.log("=========>" + err);
+        res.status(500).send("err in " + err)
+    }
+
+})
+
+
+router.delete("/:id", async (req, res) => {
+
+    const story = db.collection("story")
+
+    const token = req.headers.token
+    req.user = null;
+
+
+    if (token) {
+        data = jwt.verify(token, process.env.secritkey)
+        req.user = data
+    } else {
+        return res.status(400).json({ messege: "login frist" })
+    }
+
+    try {
+
+        livestory = await story.findOne({ "_id": new ObjectId(req.params.id) })
+
+        if (!livestory) {
+
+            return res.status(400).json({ "message": "dont find this story" })
         }
 
-        await story.deleteOne({"_id":new ObjectId(req.params.id)})
+        if (req.user.id != livestory.user) {
 
-        res.status(200).json({"message": "story  deleted"})
+            return res.status(400).json({ message: "dont allowed" })
+        }
 
-    }catch (err) {
+        if (livestory.img.publicid != null) {
+
+            cloud_remove(livestory.img.publicid)
+        }
+
+        await story.deleteOne({ "_id": new ObjectId(req.params.id) })
+
+        res.status(200).json({ "message": "story  deleted" })
+
+    } catch (err) {
         console.log("=========>" + err);
         res.status(500).send("err in " + err)
     }
